edit1

+++ src/main/java/com/envio/envio/integration/ChilexpressApiService.java
@@ -5,21 +5,21 @@ import org.springframework.stereotype.Service;
 @Service
 public class ChilexpressApiService {
 
-    public boolean notificarNuevoEnvio(Integer idEnvio) {
+    public boolean notificarNuevoEnvio(Long idEnvio) {
         // Aquí iría la lógica real para llamar a la API de Chilexpress
         System.out.println("Chilexpress API: Notificando nuevo envío con ID: " + idEnvio);
         // Simulación de llamada exitosa
         return true;
     }
 
-    public boolean actualizarEstadoEnvio(Integer idEnvio, String nuevoEstado) {
+    public boolean actualizarEstadoEnvio(Long idEnvio, String nuevoEstado) {
         // Aquí iría la lógica real para llamar a la API de Chilexpress
         System.out.println("Chilexpress API: Actualizando estado del envío " + idEnvio + " a: " + nuevoEstado);
         // Simulación de llamada exitosa
         return true;
     }
 
-    public boolean cancelarEnvio(Integer idEnvio) {
+    public boolean cancelarEnvio(Long idEnvio) {
         // Aquí iría la lógica real para llamar a la API de Chilexpress
         System.out.println("Chilexpress API: Cancelando envío con ID: " + idEnvio);
         // Simulación de llamada exitosa
